refactor(modal): migrate window-scope callers to unified modal (stage 2)
Stage 2 of SPEC_UNIFIED_MODAL_SYSTEM. Swaps the 8 `@/element/modal-v2`
importers to `@/element/modal`. A pure import-path change — `window` is
the unified primitive's default scope and the API is identical, so
behaviour is unchanged for every one of these dialogs (About, command
palette, message/user-input modals, token-breakdown popover, agent
view + action bar, import-preview).

modal-v2.* still in place for now; stage 3 migrates the tab-scope agent
modals, stage 5 deletes modal-v2.
