+++ .changesets/1779431857-refactor-modal-migrate-window-scope-callers-to-unified-modal-zef5.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(modal): migrate window-scope callers to unified modal (stage 2)
