+++ frontend/app/view/agent/components/AgentActionBar.tsx
@@ -5,7 +5,7 @@ import { createSignal, Show, type JSX } from "solid-js";
 import { RpcApi } from "@/app/store/rpc-api";
 import { TabRpcClient } from "@/app/store/rpc-util";
 import { Button } from "@/element/button";
-import { Modal, ModalBody, ModalFooter, ModalHeader } from "@/element/modal-v2";
+import { Modal, ModalBody, ModalFooter, ModalHeader } from "@/element/modal";
 import { ImportPreviewModal } from "./ImportPreviewModal";
 
 export const AgentActionBar = (): JSX.Element => {
